requirements.txt
playwright>=1.52.0
PyYAML>=6.0.1
